Move plot_ecm to visualization module
